Removes unused files; restructures src directory
* Plugins are outdated and have been removed in favor of the aria-components packages from NPM
* Examples removed, since the new configuration setup is less complicated
* Legacy Component and ComponentManager in favor of componentProvider and componentLoader
* limiter removed because we're  no longer using the manifest pattern
* fastdom dependency removed as unused (it was only used by the "plugins")

In addition, after removing the plugins directory, the src/core/ directory structure no longer makes sense
