+++ src/componentLoader.js
