--- src/core/index.js
@@ -1,7 +1,0 @@
-import Component from './Component';
-import ComponentManager from './ComponentManager';
-
-export {
-  ComponentManager,
-  Component,
-};
