--- src/plugins/index.js
@@ -1,11 +1,0 @@
-import Aria from './Aria';
-import AriaTablist from './AriaTablist';
-import AriaPopup from './AriaPopup';
-import AriaDialog from './AriaDialog';
-
-export {
-  Aria,
-  AriaTablist,
-  AriaPopup,
-  AriaDialog,
-};
